OMSA/requirements.txt
asgiref==3.8.1
attrs==25.3.0
Django==5.2.1
django-ckeditor==6.3.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
mssql-django==1.6
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
PyJWT==2.9.0
pyodbc==5.2.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
whitenoise==6.11.0
